Rename packages, fix incorrect unpack for linux and macos, add support unpack xz archives
