+++ plugins/hulylabs.langconfigurator/hulylabs.langconfigurator.iml
@@ -3,8 +3,8 @@
   <component name="NewModuleRootManager" inherit-compiler-output="true">
     <exclude-output />
     <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/resources" type="java-resource" />
+      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
@@ -12,8 +12,29 @@
     <orderEntry type="module" module-name="intellij.platform.ide.impl" />
     <orderEntry type="module" module-name="intellij.platform.ide.util.io" />
     <orderEntry type="module" module-name="intellij.platform.execution.impl" />
+    <orderEntry type="module" module-name="intellij.platform.analysis.impl" />
+    <orderEntry type="module" module-name="intellij.platform.util" />
     <orderEntry type="module" module-name="redhat.lsp4ij" />
-    <orderEntry type="library" name="slf4j-api" level="project" />
+    <orderEntry type="module-library">
+      <library name="tukaani.xz" type="repository">
+        <properties include-transitive-deps="false" maven-id="org.tukaani:xz:1.10">
+          <verification>
+            <artifact url="file://$MAVEN_REPOSITORY$/org/tukaani/xz/1.10/xz-1.10.jar">
+              <sha256sum>95c63c1a55b22dd6453890a419cc1a640f790bbf7d8ae82db1e30aefefb08888</sha256sum>
+            </artifact>
+          </verification>
+        </properties>
+        <CLASSES>
+          <root url="jar://$MAVEN_REPOSITORY$/org/tukaani/xz/1.10/xz-1.10.jar!/" />
+        </CLASSES>
+        <JAVADOC>
+          <root url="jar://$MAVEN_REPOSITORY$/org/tukaani/xz/1.10/xz-1.10-javadoc.jar!/" />
+        </JAVADOC>
+        <SOURCES>
+          <root url="jar://$MAVEN_REPOSITORY$/org/tukaani/xz/1.10/xz-1.10-sources.jar!/" />
+        </SOURCES>
+      </library>
+    </orderEntry>
     <orderEntry type="module-library">
       <library name="esotericsoftware.yamlbeans" type="repository">
         <properties include-transitive-deps="false" maven-id="com.esotericsoftware.yamlbeans:yamlbeans:1.17">
@@ -34,7 +55,5 @@
         </SOURCES>
       </library>
     </orderEntry>
-    <orderEntry type="module" module-name="intellij.platform.util" />
-    <orderEntry type="module" module-name="intellij.platform.analysis.impl" />
   </component>
 </module>
