+++ plugins/hulylabs.langconfigurator/resources/lsp-templates/rust-analyzer/clientSettings.json
