+++ plugins/hulylabs.langconfigurator/resources/lsp-templates/rust-analyzer/initializationOptions.json
