+++ plugins/hulylabs.langconfigurator/resources/lsp-templates/rust-analyzer/template.yaml
@@ -8,10 +8,12 @@ programArgs:
 binaryUrls:
   windows-x86_64: "https://github.com/rust-lang/rust-analyzer/releases/download/2024-12-09/rust-analyzer-x86_64-pc-windows-msvc.zip"
   windows-aarch64: "https://github.com/rust-lang/rust-analyzer/releases/download/2024-12-09/rust-analyzer-aarch64-pc-windows-msvc.zip"
-  unix-x86_64: "https://github.com/rust-lang/rust-analyzer/releases/download/2024-12-09/rust-analyzer-x86_64-unknown-linux-gnu.gz"
-  unix-aarch64: "https://github.com/rust-lang/rust-analyzer/releases/download/2024-12-09/rust-analyzer-aarch64-unknown-linux-gnu.gz"
+  linux-x86_64: "https://github.com/rust-lang/rust-analyzer/releases/download/2024-12-09/rust-analyzer-x86_64-unknown-linux-gnu.gz"
+  linux-aarch64: "https://github.com/rust-lang/rust-analyzer/releases/download/2024-12-09/rust-analyzer-aarch64-unknown-linux-gnu.gz"
   mac-x86_64: "https://github.com/rust-lang/rust-analyzer/releases/download/2024-12-09/rust-analyzer-x86_64-apple-darwin.gz"
   mac-aarch64: "https://github.com/rust-lang/rust-analyzer/releases/download/2024-12-09/rust-analyzer-aarch64-apple-darwin.gz"
+
+binaryExecutable: "rust-analyzer"
 
 mappingSettings:
   - languageId: "rust"
