+++ plugins/hulylabs.langconfigurator/resources/lsp-templates/typescript-language-server/clientSettings.json
