+++ plugins/hulylabs.langconfigurator/resources/lsp-templates/typescript-language-server/settings.json
