+++ plugins/hulylabs.langconfigurator/resources/lsp-templates/typescript-language-server/template.yaml
@@ -21,5 +21,5 @@ mappingSettings:
       patterns: ["*.ts"]
   - languageId: typescriptreact
     fileType:
-      name: ypeScript-React
+      name: TypeScript-React
       patterns: ["*.tsx"]
