+++ plugins/hulylabs.langconfigurator/resources/messages/HulyLangConfiguratorBundle.properties
