+++ plugins/hulylabs.langconfigurator/src/com/hulylabs/intellij/plugins/langconfigurator/LanguageServerConfigSetupNotificationProvider.java
@@ -1,11 +1,14 @@
 // Copyright © 2024 HulyLabs. Use of this source code is governed by the Apache 2.0 license.
-package hulylabs.hulycode.plugins.hulylangconfigurator;
+package com.hulylabs.intellij.plugins.langconfigurator;
 
+import com.hulylabs.intellij.plugins.langconfigurator.templates.HulyLanguageServerTemplate;
+import com.hulylabs.intellij.plugins.langconfigurator.templates.HulyLanguageServerTemplateManager;
 import com.intellij.codeInsight.daemon.DaemonCodeAnalyzer;
 import com.intellij.ide.util.PropertiesComponent;
 import com.intellij.notification.NotificationGroup;
 import com.intellij.notification.NotificationGroupManager;
 import com.intellij.notification.NotificationType;
+import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.fileEditor.FileEditor;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
@@ -15,8 +18,6 @@ import com.intellij.ui.EditorNotificationPanel;
 import com.intellij.ui.EditorNotificationProvider;
 import com.intellij.ui.EditorNotifications;
 import com.redhat.devtools.lsp4ij.LanguageServersRegistry;
-import hulylabs.hulycode.plugins.hulylangconfigurator.templates.HulyLanguageServerTemplate;
-import hulylabs.hulycode.plugins.hulylangconfigurator.templates.HulyLanguageServerTemplateManager;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -24,7 +25,7 @@ import javax.swing.*;
 import java.util.Locale;
 import java.util.function.Function;
 
-import static hulylabs.hulycode.plugins.hulylangconfigurator.messages.HulyLangConfiguratorBundle.message;
+import static com.hulylabs.intellij.plugins.langconfigurator.messages.HulyLangConfiguratorBundle.message;
 
 
 public class LanguageServerConfigSetupNotificationProvider implements EditorNotificationProvider {
@@ -35,7 +36,11 @@ public class LanguageServerConfigSetupNotificationProvider implements EditorNoti
                                                                                                                  @NotNull VirtualFile file) {
     String ext = file.getExtension() != null ? file.getExtension().toLowerCase(Locale.ROOT) : "";
     if (PropertiesComponent.getInstance(project).getBoolean(IGNORED_EXTS_KEY_PREFIX + ext)) return null;
-    HulyLanguageServerTemplate template = HulyLanguageServerTemplateManager.getInstance().getTemplate(project, file);
+
+    HulyLanguageServerTemplateManager manager = ApplicationManager.getApplication().getService(HulyLanguageServerTemplateManager.class);
+    if (manager == null) return null;
+
+    HulyLanguageServerTemplate template = manager.getTemplate(file);
     if (template == null) return null;
 
     if (LanguageServersRegistry.getInstance().isFileSupported(file, project)) return null;
