+++ plugins/hulylabs.langconfigurator/src/com/hulylabs/intellij/plugins/langconfigurator/messages/HulyLangConfiguratorBundle.java
@@ -1,13 +1,11 @@
 // Copyright 2000-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
-package hulylabs.hulycode.plugins.hulylangconfigurator.messages;
+package com.hulylabs.intellij.plugins.langconfigurator.messages;
 
 import com.intellij.DynamicBundle;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.PropertyKey;
-
-import java.util.function.Supplier;
 
 public final class HulyLangConfiguratorBundle {
   private static final @NonNls String BUNDLE = "messages.HulyLangConfiguratorBundle";
